refactor: separated AddLiquidity and MyLiquidity into different stateful widgets
